extract Value{,Type} to the top level of ::concept
